\name{decompr-package}
\alias{decompr-package}
\alias{decompr}
\docType{package}
\title{Export Decomposition using the Wang-Wei-Zhu algorithm}
\description{

}
\details{
\tabular{ll}{
Package: \tab ExDecompR\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2014-07-08\cr
License: \tab GPL-3 \cr
}
decomp: runs the decomposition
}
\author{
Bastiaan Quast

Maintainer: Bastiaan Quast <bquast@gmail.com>
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.
}
\keyword{ package }
\seealso{
}
\examples{
1+1
}
