% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{decompr}
\alias{decompr}
\alias{decompr-package}
\title{Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.}
\description{
Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
\references{
{Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.

Mehrotra, Rajiv, Fu K. Kung, and William I. Grosky. Industrial part recognition using a component-index. Image and Vision Computing 8, no. 3 (1990): 225-232.}
}

