% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{data}
\name{final.demand}
\alias{final.demand}
\title{World Input-Output Database 2011}
\description{
the final demand data
}
\references{
{Marcel P. Timmer (ed) (2012), "The World Input-Output Database (WIOD): Contents Sources and Methods",
WIOD Working Paper Number 10, downloadable at http://www.wiod.org/publications/papers/wiod10.pdf }
}

