% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kung.fu}
\alias{kung.fu}
\title{Kung Fu Export Source Decomposition}
\usage{
kung.fu(x)
}
\arguments{
\item{x}{ane object of class decompr}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
This function runs the decomposition
}
\details{
The Kung Fu algorithm
}
\author{
Bastiaan Quast
}

