% Generated by roxygen2 (4.1.0.9001): do not edit by hand
% Please edit documentation in R/leontief_output.R
\name{leontief_output}
\alias{leontief_output}
\title{Leontief Output Decomposition}
\usage{
leontief_output(x, long = TRUE)
}
\arguments{
\item{x}{ane object of class decompr}

\item{long}{transform the output data into a long (tidy) data set or not, default it TRUE.}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
This function runs the Leontief decomposition on output
}
\examples{
# load leather example data
data(leather)

# create intermediate object (class decompr)
decompr_object <- load_tables_vectors(inter,
                                      final,
                                      countries,
                                      industries,
                                      out        )

# run the Leontief decomposition on the decompr object
leontief_output(decompr_object)
}
\author{
Bastiaan Quast
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
}

