% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.tables}
\alias{load.tables}
\title{Load the Input-Output and Final demand tables}
\usage{
load.tables(x, y)
}
\arguments{
\item{x}{the intermediate demand input-output table}

\item{y}{the final demand table}
}
\value{
a decompr class object
}
\description{
This function loads the demand tables
and defines all variables for the decomposition
}
\details{
Adapted from code by Fei Wang.
}
\author{
Bastiaan Quast
}

