% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/vertical_specialisation.R
\name{vertical_specialisation}
\alias{vertical_specialisation}
\title{Vertical Specialisation}
\usage{
vertical_specialisation(x)
}
\arguments{
\item{x}{ane object of class decompr}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
This function runs the Leontief decomposition decomposition
}
\examples{
# load leather example data
data(leather)

# create intermediate object (class decompr)
decompr_object <- load_tables_vectors(inter,
                                      final,
                                      countries,
                                      industries,
                                      out        )
str(decompr_object)

# run the Vertical Specialisation decomposition on the decompr object
vertical_specialisation(decompr_object)
}
\author{
Bastiaan Quast
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
}

