% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwz.R
\name{wwz}
\alias{wwz}
\title{Wang-Wei-Zhu Decomposition of Gross Exports}
\usage{
wwz(x, verbose = FALSE)
}
\arguments{
\item{x}{an object of the class 'decompr' obtained from \code{\link{load_tables_vectors}}.}

\item{verbose}{logical, should timings of the calculation be displayed? Default is FALSE}
}
\description{
This function performs the Wang-Wei-Zhu decomposition of country-sector level gross exports into 16 value added components by importing country.
}
\details{
Adapted from code by Fei Wang.
}
\examples{
# Load example data
data(leather)

# Create intermediate object (class 'decompr')
decompr_object <- load_tables_vectors(leather)

# Perform the WWZ decomposition
wwz(decompr_object)
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu (2013). Quantifying international production sharing at the bilateral and sector levels (No. w19677). \emph{National Bureau of Economic Research}.
}
\seealso{
\code{\link{bm}}, \code{\link{kww}}, \code{\link{wwz2kww}}, \code{\link{decompr-package}}
}
\author{
Bastiaan Quast
}
