% Generated by roxygen2 (4.0.1): do not edit by hand
\name{decomp}
\alias{decomp}
\title{Interface function for decompositions}
\usage{
decomp(intermediate.demand, final.demand, method = "wwz")
}
\arguments{
\item{input.file}{the file which is
created by the load.demand function}

\item{method}{user specifies the decomposition method}
}
\value{
the decomposition table
}
\description{
This function runs the decomposition
}
\author{
Bastiaan Quast
}

