% Generated by roxygen2 (4.0.2): do not edit by hand
\name{decomp}
\alias{decomp}
\title{Interface function for decompositions}
\usage{
decomp(x, y, k, i, o, method = c("wwz", "leontief", "source"))
}
\arguments{
\item{x}{the intermediate demand table, it has dimensions GN x GN (G = no. of country, N = no. of industries),
excluding the first row and the first column which contains the country names,
and the second row and second column which contain the industry names for each country.
In addition, an extra row at the end should contain final demand.}

\item{y}{the final demand table it has dimensions GN x MN,
excluding the first row and the first column which contains the country names,
the second column which contains the industry names for each country,
and second row which contains the five decomposed final demands (M).
#'}

\item{k}{is a vector of country of region names}

\item{i}{is a vector of sector or industry names}

\item{o}{is a vecotr of final outputs}

\item{method}{user specifies the decomposition method}
}
\value{
The output when using the WWZ algorithm is a matrix with dimensions GNG*19.
 Whereby 19 is the 16 objects the WWZ algorithm decomposes exports into, plus three checksums.
 GNG represents source country, using industry and using country.
}
\description{
This function runs the decomposition
}
\examples{
# load World Input-Output Database for 2011
data(oil)

# explore the data
inter
final
countries
industries
out

# use the direct approach
# run the WWZ decomposition
wwz <- decomp(inter,
             final,
             countries,
             industries,
             out,
             method = "wwz")
wwz[1:5,1:5]

# run the Leontief decomposition
leontief  <- decomp(inter,
                    final,
                    countries,
                    industries,
                    out,
                    method = "leontief")
leontief
}
\author{
Bastiaan Quast
}
\references{
{Timmer, Marcel P. (ed) (2012), "The World Input-Output Database (WIOD): Contents Sources and Methods", WIOD Working Paper Number 10, downloadable at http://www.wiod.org/publications/papers/wiod10.pdf }

{Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.}
}

