% Generated by roxygen2 (4.0.1): do not edit by hand
\name{decomp}
\alias{decomp}
\title{Interface function for decompositions}
\usage{
decomp(x, y, method = "wwz")
}
\arguments{
\item{x}{the intermediate demand table, where the first row and the first column list the country names, the second row and second column list the insdustry names for each country. The matrix is presumed to be of dimensions G*N where G represents the country and N the industry. No extra columns should be there. Extra rows at the bottom which list adjustments such as taxes as well are disregarded, with the exception that the very last row is presumed to contain the total output.}

\item{y}{the final demand table it has dimensions GN*GM ( where M is the number of objects final demand is decomposed in, e.g. household consumption, here this is five decompositions).}

\item{method}{user specifies the decomposition method}
}
\value{
The output when using the WWZ algorithm is a matrix with dimensions GNG*19. Whereby 19 is the 16 objects the WWZ algorithm decomposes exports into, plus three checksums. GNG represents source country, using industry and using country.
}
\description{
This function runs the decomposition
}
\author{
Bastiaan Quast
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
Mehrotra, Rajiv, Fu K. Kung, and William I. Grosky.
"Industrial part recognition using a component-index."
Image and Vision Computing 8, no. 3 (1990): 225-232.
}

