\name{decompr-package}
\alias{decompr-package}
\alias{decompr}
\docType{package}
\title{Export Decomposition }
\description{
Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.
}
\details{
\tabular{ll}{
Package: \tab decompr \cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2014-07-14\cr
License: \tab GPL-3 \cr
}
decomp: runs the decomposition
load.demand: creates the decompr object which can be used by the individual methods
wwz: takes a decompr object and performs the Wang-Wei-Zhu decomposition
kung.fu takes a decompr object and performs the Kung-Fu decomposition
}
\author{
Bastiaan Quast

Maintainer: Bastiaan Quast <bquast@gmail.com>
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.
}
\keyword{ package }
\seealso{
}
\examples{
1+1
}
