\name{decompr-package}
\alias{decompr-package}
\alias{decompr}
\docType{package}
\title{Export Decomposition }
\description{
Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.
}
\details{
\tabular{ll}{
Package: \tab decompr \cr
Type: \tab Package\cr
Version: \tab 0.3.0\cr
Date: \tab 2014-07-14\cr
License: \tab GPL-3 \cr
}
decomp: runs the decomposition
load.tables: creates the decompr object which can be used by the individual methods
wwz: takes a decompr object and performs the Wang-Wei-Zhu decomposition
kung.fu takes a decompr object and performs the Kung-Fu decomposition
}
\author{
Bastiaan Quast

Maintainer: Bastiaan Quast <bquast@gmail.com>
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.

Mehrotra, Rajiv, Fu K. Kung, and William I. Grosky. "Industrial part recognition using a component-index." Image and Vision Computing 8, no. 3 (1990): 225-232.
}
\keyword{ package }
