% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{decompr}
\alias{decompr}
\alias{decompr-package}
\title{Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.}
\description{
Export Decomposition using the Wang-Wei-Zhu and Kung-Fu algorithms.
}
\examples{
# load World Input-Output Database for 2011
data(wiod)

# explore the data
dim(intermediate.demand) # (2 + GN + totals) x (2 + GN)
dim(final.demand)        # (2 + GN + totals) x (2 + G*5)
intermediate.demand[1:40,1:40]
final.demand[1:40,1:10]

# use the direct approach
# run the WWZ decomposition
wwz <- decomp(intermediate.demand, final.demand, method='wwz')
wwz[1:5,1:5]

# run the Kung Fu decomposition
kf  <- decomp(intermediate.demand, final.demand, method='kung.fu')
kf[1:5,1:5]
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
}
\references{
{Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.

Mehrotra, Rajiv, Fu K. Kung, and William I. Grosky. Industrial part recognition using a component-index. Image and Vision Computing 8, no. 3 (1990): 225-232.}
}

