% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decompr.R
\docType{package}
\name{decompr}
\alias{decompr}
\alias{decompr-package}
\title{Export Decomposition using the Wang-Wei-Zhu and Leontief decompositions algorithms.}
\description{
Export Decomposition using the Wang-Wei-Zhu and Leontief decompositions algorithms.
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
Fei Wang
Victor Kummritz
}
\references{
{Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.}
}

