% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decompr.R
\docType{package}
\name{decompr}
\alias{decompr}
\alias{decompr-package}
\title{Export Decomposition using the Wang-Wei-Zhu and Leontief decompositions algorithms.}
\description{
Two global-value-chain decompositions are implemented. Firstly, the
Wang-Wei-Zhu (Wang, Wei, and Zhu, 2013) algorithm splits bilateral gross exports
into 16 value-added components. Secondly, the Leontief decomposition (default)
derives the value added origin of exports by country and industry, which is also
based on Wang, Wei, and Zhu (Wang, Z., S.-J. Wei, and K. Zhu. 2013. "Quantifying
International Production Sharing at the Bilateral and Sector Levels.").
}
\references{
{Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu. Quantifying international production sharing at the bilateral and sector levels. No. w19677. National Bureau of Economic Research, 2013.}
}
\seealso{
http://qua.st/decompr
}
\author{
Bastiaan Quast \email{bquast@gmail.com}
Fei Wang
Victor Kummritz
}
