% Generated by roxygen2 (4.0.1): do not edit by hand
\name{kung.fu}
\alias{kung.fu}
\title{Kung Fu Export Source Decomposition}
\usage{
kung.fu(x)
}
\arguments{
\item{x}{ane object of class decompr}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
This function runs the Kung Fu decomposition
}
\examples{
# load World Input-Output Database for 2011
data(wiod)

# create intermediate object (class decompr)
decompr.object <- load.tables(intermediate.demand, final.demand)
str(decompr.object)

# run the Kung Fu decomposition on the decompr object
kf  <- kung.fu(decompr.object)
kf[1:5,1:5]
}
\author{
Bastiaan Quast
}
\references{
No. w19677. National Bureau of Economic Research, 2013.
Mehrotra, Rajiv, Fu K. Kung, and William I. Grosky.
Industrial part recognition using a component-index.
Image and Vision Computing 8, no. 3 (1990): 225-232.
}

