% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/leontief.R
\name{kung_fu}
\alias{kung_fu}
\title{Source Decomposition (deprecated)}
\usage{
kung_fu(x)
}
\arguments{
\item{x}{ane object of class decompr}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
DEPRECATED: This function runs the source decomposition decomposition
}
\examples{
# load World Input-Output Database for 2011
data(wiod)

# create intermediate object (class decompr)
decompr_object <- load_tables_vectors(intermediate_demand,
                                      final_demand,
                                      region_names,
                                      industry_names,
                                      output              )
str(decompr_object)

# run the Kung Fu decomposition on the decompr object
kf  <- kung_fu(decompr_object)
kf[1:5,1:5]
}
\author{
Bastiaan Quast
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
}

