% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/leontief.R
\name{leontief}
\alias{leontief}
\title{Leontief Decomposition}
\usage{
leontief(x, post = c("exports", "output", "final_demand", "none"),
  long = TRUE)
}
\arguments{
\item{x}{an object of class decompr}

\item{post}{post-multiply the Leontief inverse with something, the default is exports}

\item{long}{transform the output data into a long (tidy) data set or not, default it TRUE.}
}
\value{
a data frame containing the square matrix and labelled column and rows
}
\description{
Leontief Decomposition
}
\examples{
## load example data
data(leather)

## create intermediate object (class decompr)
decompr_object <- load_tables_vectors(inter,
                                      final,
                                      countries,
                                      industries,
                                      out        )

## run the Leontief decomposition on the decompr object
leontief(decompr_object )
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
}
\author{
Bastiaan Quast
}
