% Generated by roxygen2 (4.0.1): do not edit by hand
\name{load.demand}
\alias{load.demand}
\title{Load the demand data}
\usage{
load.demand(IOA, WFD)
}
\arguments{
\item{rownam}{list of country and industry names,
the dimensions have to beG*N by 2.}

\item{IOA}{the intermediate demand input-ouput table}

\item{WFD}{the final demand table}
}
\value{
a exdecompr class object
the data is saved to a file (for now)
}
\description{
This function load the demand tables
and defines all variables for the decomposition
}
\details{
This function load the data
and creates the variables for the decomposition.
Adapted from code by Fei Wang.
}
\author{
Bastiaan Quast
}

