% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/decompr.R
\docType{data}
\name{region_names}
\alias{region_names}
\title{World Input-Output Database 2011}
\description{
the names of the regions data
}
\references{
{Timmer, Marcel P. (ed) (2012), "The World Input-Output Database (WIOD): Contents Sources and Methods",
WIOD Working Paper Number 10, downloadable at http://www.wiod.org/publications/papers/wiod10.pdf }
}

