% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wwz}
\alias{wwz}
\title{Runs the WWZ decomposition}
\usage{
wwz(input.object, return.totals = FALSE)
}
\arguments{
\item{input.object}{an object of the class decompr}

\item{return.totals}{logical as to whether totals should be included in the output}
}
\value{
the decomposed table
}
\description{
This function runs the decomposition
}
\details{
Adapted from code by Fei Wang.
}
\author{
Bastiaan Quast
}

