% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wwz}
\alias{wwz}
\title{Runs the WWZ decomposition}
\usage{
wwz(x)
}
\arguments{
\item{x}{an object of the class decompr}
}
\value{
the decomposed table
}
\description{
This function runs the decomposition
}
\details{
Adapted from code by Fei Wang.
}
\author{
Bastiaan Quast
}

