% Generated by roxygen2 (4.0.1): do not edit by hand
\name{wwz}
\alias{wwz}
\title{Runs the Wang-Wei-Zhu decomposition}
\usage{
wwz(x)
}
\arguments{
\item{x}{an object of the class decompr}
}
\value{
the decomposed table
}
\description{
This function runs the Wang-Wei-Zhu decomposition.
}
\details{
Adapted from code by Fei Wang.
}
\examples{
# load World Input-Output Database for 2011
data(wiod)

# create intermediate object (class decompr)
decompr_object <- load_tables(intermediate_demand, final_demand)
str(decompr_object)

# run the WWZ decomposition on the decompr object
wwz  <- wwz(decompr_object)
wwz[1:5,1:5]
}
\author{
Bastiaan Quast
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels.
No. w19677. National Bureau of Economic Research, 2013.
}

