% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wwz.R
\name{wwz}
\alias{wwz}
\title{Runs the Wang-Wei-Zhu decomposition}
\usage{
wwz(x, verbose = FALSE)
}
\arguments{
\item{x}{an object of the class decompr}

\item{verbose}{logical, should timings of the calculation be displayed? Default is FALSE}
}
\value{
the decomposed table
}
\description{
This function runs the Wang-Wei-Zhu decomposition.
}
\details{
Adapted from code by Fei Wang.
}
\examples{
# load example data
data(leather)

# create intermediate object (class decompr)
decompr_object <- load_tables_vectors(inter,
                                      final,
                                      countries,
                                      industries,
                                      out        )

# run the WWZ decomposition on the decompr object
wwz(decompr_object)
}
\references{
Wang, Zhi, Shang-Jin Wei, and Kunfu Zhu.
Quantifying international production sharing at the bilateral and sector levels. 
No. w19677. National Bureau of Economic Research, 2013.
}
\author{
Bastiaan Quast
}
